+++ test/test_doc_types.py
@@ -48,3 +48,14 @@ def test_eval_fail3(monkeypatch):
             safe_eval_type("FooFn[int]", {"typing": typing}, types.ModuleType("a"), "a")
             == "FooFn[int]"
         )
+
+
+def test_eval_union_types_on_old_python(monkeypatch):
+    monkeypatch.setattr(sys, "version_info", (3, 9, 0))
+    with pytest.warns(
+        UserWarning,
+        match=r"You are likely attempting to use Python 3.10 syntax \(PEP 604 union types\) "
+        r"with an older Python release.",
+    ):
+        # str never implements `|`, so we can use that to trigger the error on newer versions.
+        safe_eval_type('"foo" | "bar"', {}, None, "example")